+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - 2020-03-08
+### Changed
+- The integration tests were refactored and expanded to cover more use cases.
 
 ## [0.2.2] - 2020-03-07
 ### Changed